chore: move test scripts to scripts/
